+++ application/src/main/java/com/ericsson/bss/cassandra/ecchronos/application/config/Config.java
@@ -86,6 +86,7 @@ public class Config
         if (statisticsConfig != null)
         {
             myStatisticsConfig = statisticsConfig;
+            myStatisticsConfig.validate();
         }
     }