+++ application/src/test/java/com/ericsson/bss/cassandra/ecchronos/application/config/TestConfig.java
@@ -113,6 +113,7 @@ public class TestConfig
                 .withBackoff(13, TimeUnit.SECONDS)
                 .withTargetRepairSizeInBytes(UnitConverter.toBytes("5m"))
                 .withPriorityGranularityUnit(TimeUnit.MINUTES)
+                .withInitialDelay(1, TimeUnit.HOURS)
                 .build();
 
         GlobalRepairConfig repairConfig = config.getRepairConfig();
@@ -159,6 +160,10 @@ public class TestConfig
         assertThat(httpReportingConfig.getExcludedMetrics()).hasSize(1);
         assertThat(httpReportingConfig.getExcludedMetrics()).contains(expectedHttpExcludedMetric);
 
+        assertThat(statisticsConfig.getRepairFailuresTimeWindow().getInterval(TimeUnit.MINUTES)).isEqualTo(5);
+        assertThat(statisticsConfig.getTriggerIntervalForMetricInspection().getInterval(TimeUnit.SECONDS)).isEqualTo(30);
+        assertThat(statisticsConfig.getRepairFailuresCount()).isEqualTo(5);
+
         LockFactoryConfig lockFactoryConfig = config.getLockFactory();
         assertThat(lockFactoryConfig.getCasLockFactoryConfig().getKeyspaceName()).isEqualTo("ecc");
         assertThat(lockFactoryConfig.getCasLockFactoryConfig().getConsistencySerial().equals(ConsistencyType.LOCAL)).isTrue();
@@ -172,6 +177,36 @@ public class TestConfig
         RestServerConfig restServerConfig = config.getRestServer();
         assertThat(restServerConfig.getHost()).isEqualTo("127.0.0.2");
         assertThat(restServerConfig.getPort()).isEqualTo(8081);
+    }
+
+    @Test
+    public void testInitialDelayDefaultValue() throws Exception
+    {
+        ClassLoader classLoader = Thread.currentThread().getContextClassLoader();
+        File file = new File(classLoader.getResource("nothing_set.yml").getFile());
+
+        ObjectMapper objectMapper = new ObjectMapper(new YAMLFactory());
+
+        Config config = objectMapper.readValue(file, Config.class);
+        GlobalRepairConfig repair = config.getRepairConfig();
+        Interval interval = repair.getInitialDelay();
+        assertThat(interval.getInterval(TimeUnit.DAYS)).isEqualTo(1);
+    }
+
+    @Test
+    public void testInitialDelayLongerThanRepairInterval() throws Exception
+    {
+        ClassLoader classLoader = Thread.currentThread().getContextClassLoader();
+        File file = new File(classLoader.getResource("initial_delay_greater_than_repair.yml").getFile());
+
+        ObjectMapper objectMapper = new ObjectMapper(new YAMLFactory());
+        Config config = objectMapper.readValue(file, Config.class);
+        GlobalRepairConfig repair = config.getRepairConfig();
+        Interval repint = repair.getRepairInterval();
+        Interval intdel = repair.getInitialDelay();
+
+        assertThat(repint.getTime()).isEqualTo(intdel.getTime());
+        assertThat(repint.getUnit()).isEqualTo(intdel.getUnit());
     }
 
     @Test
@@ -229,6 +264,7 @@ public class TestConfig
         assertThat(repairConfig.getIgnoreTWCSTables()).isFalse();
         assertThat(repairConfig.getBackoff().getInterval(TimeUnit.MINUTES)).isEqualTo(30);
         assertThat(repairConfig.getPriority().getPriorityGranularityUnit()).isEqualTo(TimeUnit.HOURS);
+        assertThat(repairConfig.getInitialDelay().getInterval(TimeUnit.DAYS)).isEqualTo(1);
 
         StatisticsConfig statisticsConfig = config.getStatisticsConfig();
         assertThat(statisticsConfig.isEnabled()).isTrue();
@@ -338,7 +374,11 @@ public class TestConfig
         assertThat(httpReportingConfig.isEnabled()).isTrue();
         assertThat(httpReportingConfig.getExcludedMetrics()).isEmpty();
 
-        LockFactoryConfig lockFactoryConfig = config.getLockFactory();
+        assertThat(statisticsConfig.getRepairFailuresTimeWindow().getInterval(TimeUnit.MINUTES)).isEqualTo(30);
+        assertThat(statisticsConfig.getTriggerIntervalForMetricInspection().getInterval(TimeUnit.SECONDS)).isEqualTo(5);
+        assertThat(statisticsConfig.getRepairFailuresCount()).isEqualTo(5);
+
+       LockFactoryConfig lockFactoryConfig = config.getLockFactory();
         assertThat(lockFactoryConfig.getCasLockFactoryConfig().getKeyspaceName()).isEqualTo("ecchronos");
         assertThat(lockFactoryConfig.getCasLockFactoryConfig().getConsistencySerial().equals(ConsistencyType.DEFAULT)).isTrue();
 
@@ -376,7 +416,7 @@ public class TestConfig
     }
 
     @Test
-    public void testStatisticsDisabledIfNoReporting() throws Exception
+    public void testStatisticsEnabledIfNoReporting() throws Exception
     {
         ClassLoader classLoader = Thread.currentThread().getContextClassLoader();
         File file = new File(classLoader.getResource("all_reporting_disabled.yml").getFile());
@@ -389,7 +429,18 @@ public class TestConfig
         assertThat(statisticsConfig.getReportingConfigs().isJmxReportingEnabled()).isFalse();
         assertThat(statisticsConfig.getReportingConfigs().isFileReportingEnabled()).isFalse();
         assertThat(statisticsConfig.getReportingConfigs().isHttpReportingEnabled()).isFalse();
-        assertThat(statisticsConfig.isEnabled()).isFalse();
+        assertThat(statisticsConfig.isEnabled()).isTrue();
+    }
+
+    @Test
+    public void testTriggerIntervalBiggerThanRepairFailuresWindow()
+    {
+        ClassLoader classLoader = Thread.currentThread().getContextClassLoader();
+        File file = new File(classLoader.getResource("trigger_interval_bigger_than_repair_failures_window.yml").getFile());
+
+        ObjectMapper objectMapper = new ObjectMapper(new YAMLFactory());
+
+        assertThatExceptionOfType(JsonMappingException.class).isThrownBy(() -> objectMapper.readValue(file, Config.class));
     }
 
     public static class TestNativeConnectionProvider implements NativeConnectionProvider