+++ CHANGES.md
@@ -2,7 +2,17 @@
 ## Version 5.0.5
 
 * Define logging levels formally - Issue #666
+* Metric status logger for troubleshooting - Issue #397
 
+## Version 5.0.4
+
+* ecChronos will break if repair interval is shorter than the initial delay - Issue #667
+
+## Version 5.0.3
+
+* Spring Framework URL Parsing with Host Validation is vulnerable - Issue #661
+* Possibility for repairs to never be triggered - Issue #264
+ 
 ## Version 5.0.2
 
 * Containerized ecchronos restarting tomcat when Cassandra peer is overloaded - Issue #650